core: cleanup `key.ts`
The commit cleans up `key.ts` by removing the `browser` import and instead using the proper `common` one.
The code also moves `keys.spec.ts` to the proper location (`common`) and organizes it better.
