+++ packages/core/src/browser/keyboard/browser-keyboard-layout-provider.spec.ts
@@ -28,7 +28,7 @@ import { LocalStorageService } from '../storage-service';
 import { MessageService } from '../../common/message-service';
 import { WindowService } from '../window/window-service';
 import { BrowserKeyboardLayoutProvider } from './browser-keyboard-layout-provider';
-import { Key } from './keys';
+import { Key } from '../../common/keys';
 
 disableJSDOM();
