+++ packages/terminal/src/browser/search/terminal-search-widget.tsx
@@ -20,7 +20,8 @@ import * as React from '@theia/core/shared/react';
 import '../../../src/browser/style/terminal-search.css';
 import { Terminal } from 'xterm';
 import { SearchAddon, ISearchOptions } from 'xterm-addon-search';
-import { codicon, Key } from '@theia/core/lib/browser';
+import { codicon } from '@theia/core/lib/browser';
+import { Key } from '@theia/core/lib/common/keys';
 
 export const TERMINAL_SEARCH_WIDGET_FACTORY_ID = 'terminal-search';
 export const TerminalSearchWidgetFactory = Symbol('TerminalSearchWidgetFactory');
